docs(example): move game example to game dir
