+++ example/game/src/Run.ts
@@ -1,12 +1,14 @@
+/* tslint:disable: no-import-side-effect ordered-imports no-console */
+
 /**
  * Created by tushar on 2019-05-05
  */
 
-import {prompt} from 'promptly'
-// tslint:disable-next-line: no-import-side-effect
 import 'source-map-support/register'
 
-import {defaultRuntime} from '../../src/runtimes/DefaultRuntime'
+import {prompt} from 'promptly'
+
+import {defaultRuntime} from '../../../src/runtimes/DefaultRuntime'
 
 import {program} from './Program'
 
@@ -14,7 +16,6 @@ defaultRuntime().execute(
   program.provide({
     console: {
       getStrLn: prompt,
-      // tslint:disable-next-line:no-console
       putStrLn: (message: string) => console.log(message)
     },
     random: {
